Modified the way the text filter handles "exact" string matching
UE-16745 - Cannot search for assets that have a dash (-) in the name

Previously, putting a string inside quotes would result in it performing an exact string match, however this left you with no way to perform a partial string match against a string containing special characters (such as -).

I've changed the behavior of the text filter syntax so that a + (pre-unary) denotes that a string should perform an exact match, and this can be applied to both quoted and unquoted strings (which now both perform partial matches by default). This matches the old behavior of the text filter.

[CL 2582491 by Jamie Dale in Main branch]