Improved path searching in the Content Browser
Re-added the exported text for an asset to the list of simple search terms. This was broken a while ago, and messed with some peoples workflow as they used to Ctrl+C the asset to get the exported text which they could later search for in the Content Browser.

This also improves the handling of the "path" key, as it can now compare against paths in any of the following forms (previously it only tested the last one):
/Path/To/Asset.Asset
/Path/To/Asset
/Path/To

[CL 2582504 by Jamie Dale in Main branch]