Fixed bad string access in SEditableText::IsAtWordStart
UE-16760 - Crash occurs using ctrl/cmd + right arrow in text fields after copying text

FString was recently changed to prevent you accessing the null terminator from within the string (the character at index Len()), and this code was still trying to access that character, and was triggering an assert.

I've added a guard to SEditableText::IsAtWordStart to prevent this, and also checked that the other string access was suitably guarded (it was).

[CL 2582609 by Jamie Dale in Main branch]