Fixed some cases where the text filter would fail to update
The text filter was stripping whitespace and storing the stripped version. This meant that a search term containing only whitespace could never be cleared again, as the filter never notified of a change in text.

The other issue was that some code was performing a pre-test to see if the text had changed, however this code was doing case-insensitive comparisons, which no longer suffice as the operators are case-sensitive.

[CL 2587373 by Jamie Dale in Main branch]