Fixed some issues using collections with SVN or Git as your SCC provider
UE-17109 - Collections cannot be created when using SVN as your SCC provider

The collection was trying to add itself to source control before the file had been created. Perforce allows this, but SVN and Git do not.

Additionally, SVN was trying to recursively commit any added parent folders when committing a file within those folders (required to avoid the commit operation failing), however that was failing to get the current SVN status of anything but the most immediate parent folder, causing the recusive check to fail.

[CL 2588523 by Jamie Dale in Main branch]