Merging using UE4-LauncherDev-To-UE4-LauncherReleases
Adding the setting of file attributes to contribute to an Installing phase of a installer. This fixes an issue where Initialising or Verifying appears to take a long time.
Also optimising the file attributes setting code by reading file attributes before attemping to change them, this has a large effect on speed as most files will be setup already, especially for a repair operation.
#jira OPP-2904: UE4 verifications display Initializing for a long time before the verification begins
#reviewedby Justin.Sargent

[CL 2594391 by Leigh Swift in Main branch]