Made clearing the text selection when a ML editable text loses focus optional
UE-17599 - Rich Text for tutorials is not working in Main

Previously SMultiLineEditableText never cleared its selection when it lost focus, however recent changes to make it behave more like SEditableText had undone that behavior. This caused issues with some of our text editors, as they rely on the selection being preserved when focus is lost so that they can style the selected text.

I've added an option to control whether focus loss should clear the current selection, and have turned it off for our text editors.

#codereview Justin.Sargent, Ben.Cosh

[CL 2600404 by Jamie Dale in Main branch]