More text filter improvements
The "-" token as been removed from the list of general NOT operator aliases, and is now only available as TextCmpInvert. This means that negative numbers will now parse correctly, but also means that we can now handle complex expressions like "name=-Thing" as a user might expect (equivalent to "name!=thing").

The expression is now evaluated once after compilation (using a dummy context) so that syntax error information can be reported back to the user. Previously evaluation error information was being lost.

#codereview Andrew.Rodham

[CL 2602053 by Jamie Dale in Main branch]