UE-17106 - Change the dynamic parameter default value back to 1.
Added DefaultValue property to DynamicParameter material node. This defaults to 1 for new materials but 0 for existing ones to maintain current behavior.
This default is used for all vertex factories other than those for particles. In particle vertex factories, the dynamic parameter is used, as before. Sometimes coming from vertex data, others coming from a default vertex buffer filled with 1.0.

Also, defaulted the DynamicParameter on the cascade side to 1.0 to improve workflow that side too.

#codereview Olaf.Piesche

[CL 2605332 by Simon Tovey in Main branch]