Made asset view tooltips refresh each time they become visible
UE-17564 - Collections are not removed from Asset tooltip when it's removed from collection

This allows the list of collections that an object is in to be updated, as this isn't stored as part of the asset data.

The main set of tags (which is the bulk of the tooltip text) is still cached and only updated when the asset data is changed.

[CL 2605509 by Jamie Dale in Main branch]