Fixed the "Show Collections" option appearing multiple times if multiple Content Browsers were open
UE-17880 - Show Collections shows up twice when box is checked and multiple content browsers

This used to be added via an extender, however the extender was bound once each time a Content Browser was opened, meaning it was possible to have the entry added multiple times if multiple Content Browsers were open.

This menu item is now handled by the SAssetView (the same as the other view options), and is also now disabled on dialogs where it doesn't make sense to change the "Show Collections" settings (eg, when using the load or save picker dialogs, as they don't contain a collections view).

[CL 2605762 by Jamie Dale in Main branch]