UE-11390 - Parameters not appearing in Material instances
This is now "fixed". This was only ever a problem for parameters that fed into a function input's Preview pin.
They will never have an impact on the final material's color and so are useless. But as it's such a minor edge case, it's not worth the effort of the real solution which would be to disallow connecting parameters to the preview in in the first place. Or something along those lines.

[CL 2608348 by Simon Tovey in Main branch]