Object Reference, Class Reference, Asset ID, Asset Class ID pin types have been spun off into a sub-menu on the item to reduce duplication of these items in the listview. Selecting the item in the main tree will auto-select object reference as the pin type for the specified Blueprint type.
Spun out management of sub-menus in SMultiBox to the new class SMenuOwner for reusability.

Fixed potential issue with table views that would prevent them from stopping to be marked as hovered.

Added SSubMenuHandler to provide duplicate functionality to SMenuEntryBlock including deferred menu opening and close, and allowing the user's mouse to cross over other sibling items in order to get to the currently opened sub-menu.

#jira UE-16951 - Move assetptr references from the type selector to an option on the variable

[CL 2609080 by Michael Schoell in Main branch]