Added stricter validation for collection names
UE-17561 - Collections allow for several special characters that are not normally allowed
UE-17549 - User gets no warning for invalid Characters when naming a Collection
UE-17538 - Add a Character limit for naming a collection

The list of valid characters now matches that of assets and folders in the Content Browser, and the user is informed when they try and use an invalid character.

The collection name is also now restricted to NAME_SIZE to avoid a crash if a user entered a name that was too long.

[CL 2609261 by Jamie Dale in Main branch]