UMG/Slate - Initial version of SInvalidationPanel!  Has several limitations, but permits UMG and Slate to effectively "Freeze" UI, caching a great deal of the processing, computation...etc required every time slate paints.  Cached widgets do not get prepass, tick, or paint until their parent SInvalidationPanel is invalidated.  Widgets that change often, can be marked as Volatile and get treated special when used in conjunction with the SInvalidationPanel, they get stored seperately and get prepass/tick/paint like normal, but using the cached geometry of their container they origionally were computed to have.  Changes to their desired size will not change the invalidation of the rest of the panel, a manual invocation would be required for that.
When used effectively the SInvalidationPanel can shave milliseconds off your Slate time on a complicated hierarchy of widgets.  Provided those widgets don't need hit testing or input, this can be a very useful widget to use.  Scenarios like a HUD are the primary target of this panel, future advances will hopefully make it possible to be used in conjunction with a UI that needs input.

#codereview matt.kuhlenschmidt, nick.atamas

[CL 2609569 by Nick Darnell in Main branch]