Added status information to collections
This shows a little colored pip to the right of the collection items, where each color means:
  Grey   - Up-to-date and empty
  Green  - Up-to-date and not empty
  Blue   - Has local changes
  Orange - Not at the latest revision
  Red    - Checked out by another user, in a conflicted state, or missing its source control provider

This also adds an "Update" and "Save" option to the collections context menu. Typically these won't be needed as collections mostly manage themselves, but since we now show when a collection is out-of-date or dirty, it makes sense to allow the user to take action to correct these states without having to leave the editor.

[CL 2610477 by Jamie Dale in Main branch]