Made the collection manager caching lazy
This used to maintain an up-to-date cache as soon as a change happened. This lead to extra complexity in some cases, and often would keep caches up-to-date that weren't currently even being used.

Adding, removing, or updating a collection will now just mark the relevant parts of the cache as dirty so that they can be updated (using a well defined, safe, full update) the next time something requests them.

[CL 2610959 by Jamie Dale in Main branch]