Optimizations to setting the material or its texture transform on a BSP surface. #jira UE-16413 - Applying materials to BSP takes a long time once shells reach a basic level of complexity
Previously, setting a material or its texture transform on a BSP surface triggered CommitModelSurfaces, which iterated all the ModelComponents in the level and caused (among other things) the physics state to be recreated, a very expensive operation.  This is clearly unnecessary when only the surface material or texture transform has changed, so introduced a new path in which only the Material Index Buffers are updated for the level's UModel.

[CL 2613582 by Richard TalbotWatkin in Main branch]