Fixed crash when shutting down the editor
UE-18403 - Crash occurs exiting the editor with Collections visible

The SCC provider would be nulled out when the SCC module is unloaded, which was triggering a notification to update the collection item states, however the collection manager module had also been unloaded by the time that was being processed, leading to an assert.

This change makes sure that the collection manager module is still available before attempting to use it to update the collection item states.

[CL 2616559 by Jamie Dale in Main branch]