Improved the folder blacklist in GameProjectUtils::DuplicateProjectForUpgrade
UE-18319 - Collections reset when the engine upgrades project

GameProjectUtils::DuplicateProjectForUpgrade used to wholesale skip the Saved directory, which meant that copying your project could appear to lose your local collections, as well as any saved user-settings.

This change allows it to copy the parts of the Saved directory that we know aren't transient. The RelativeDirectoriesToSkip array may need updating as new Saved content is added in order to keep the copy as fast as possible.

#codereview Ben.Marsh

[CL 2618570 by Jamie Dale in Main branch]