Improved the behavior of Ctrl+B in the Content Browser and Persona
Ctrl+B in the Content Browser will now invoke the "Show in Folder View" command to allow you to easily sync to the actual location of an asset in the cases where you're viewing it via a search or a collection.

Ctrl+B in the Persona animation sequence browser will now invoke the "Find in Content Browser" command.

Ctrl+B will no longer do anything when Persona has no asset available for a given edit mode.

[CL 2619988 by Jamie Dale in Main branch]