Improved the UX in the composite font editor
UE-18107 - Broken UX in Font Editor

Adding a new font or sub-font family will now automatically open their names for editing to teach you that they can be renamed.

This also changed the icon for opening a font file as there was some confusion about the ellipsis being for the text that was cut off, rather than being a button to pick a font file.

[CL 2622006 by Jamie Dale in Main branch]