Fixed the collection manager watching non-collection directories for changes
UE-18771 - Collections log spam when building engine dlls that live in the same directory as a running instance of the editor

The collection manager now ensures that the collection directories exist on disk before it starts watching them, and will also ignore the "system" collection directory (as that is blank and was causing it to watch for changes in the Binaries folder).

As a belt-and-braces measure, FCollectionManager::TickFileCache will also now ensure that a file ends with the ".collection" extension before attempting to process the file as a collection.

[CL 2627607 by Jamie Dale in Main branch]