Fixed the Reference Viewer "Make Collection with Referenced Assets" option failing
UE-18971 - ReferenceViewer_Make Collection with Referenced Asset Local Broken

This had fallen foul of the new naming restrictions for collections, and has been changed to use the leaf-name of the asset, rather than its full path.

This also adds the collection type icons to the sub-menu in the Reference Viewer.

[CL 2628794 by Jamie Dale in Main branch]