Updated SlateFileDialogs to no longer be dependent on SLATE_TEXT_ATTRIBUTE
SLATE_TEXT_ATTRIBUTE was the macro that allowed the FString -> FText pass-through support, and was deprecated in 4.8 in favour of using SLATE_ATTRIBUTE with an FText type. This is being removed for 4.10 and SlateFileDialogs was still relying on the deprecated behaviour.

This change removes its reliance on SLATE_TEXT_ATTRIBUTE, and also cleans up a load of SLATE_ATTRIBUTE parameters should have been using SLATE_ARGUMENT instead.

#codereview Dmitry.Rekman

[CL 2628829 by Jamie Dale in Main branch]