HMD hidden area mesh implementation.
Adds an interface for rendering an HMD's hidden area mesh as a depth prepass.
Either runs in the standard depth prepass if it's being used, or runs as its own prepass.
Runs only if running in stereo, the device has a hidden area mesh defined and vr.HiddenAreaMask is set to 1.

[CL 2632589 by Ryan Vance in Main branch]