Converted the UE4 Windows filesysytem implementation to use Windows API functions for getting file stat data
This yields better performance than using the POSIX wrapper functions, and will also allow us to use the stat data available while walking the directories (see IterateDirectory) without having issues with the two timestamps having different precisions.

Note: The new timestamps are more precise (they have 100-nanosecond precision), so this will result in anything that caches using only timestamps (such as the asset registry) performing a full re-scan. The auto-reimport is immune to this as it also tests the MD5 of the file.

#platformnotify Josh.Adams

[CL 2634112 by Jamie Dale in Main branch]