Updated the asset registry gathering to use the directory iteration functions that provide file stat data
The discovery phase now uses the fact that the stat data for a file is available up-front to perform the check against the cached timestamp immediately. On Windows, and for the cases where most or all of the data is cached, this can significantly improve the asset registry gathering time as it avoids performing a second scan of the file-system for the majority of the files.

[CL 2634348 by Jamie Dale in Main branch]