[AUTOMERGE]
STooltipPresenter was keeping a strong pointer to the tooltip, so in a case where it wasn't explicitly set to null it would keep around a stale tooltip. This occured when the owning widget of the tooltip was destroyed.

#RB Nick.Atamas
#codereview Nick.Atamas, Matt.Kuhlenschmidt

--------
Integrated using branch Ue4-To-UE4-Fortnite-Simple (reversed) of change#2634797 by Chris.Gagnon on 2015/07/27 19:14:39.

[CL 2634804 by Chris Gagnon in Main branch]