Changed FUnrealEdMisc::TickAssetAnalytics to use more appropriate containers
It was using two TArrays with AddUnique, which was *very* slow for large projects, and was causing a hitch after the asset registry had finished populating.

These have been changed to be a TSet and TMap, which can handle unique entries for large data sets (1,000,000 in this particular case) much more efficiently.

#codereview Ben.Cosh

[CL 2638480 by Jamie Dale in Main branch]