Improved SubsurfaceScatteringProfile quality, base color is now applied after scattering which gives crisper textures and more control.
This is more correct for scanned textures as they include scattering already. Affects the look of emissive lighting (rarely needed and can be fixed by content). All diffuse lighting for SubsurfaceScatteringProfile is now done without the BaseColor (minor performance cost) and reading the Lighting before the SubsurfaceScattering pass would be different.

Removed the need for a SceneColor alpha channel as lighting is now stored in a checkerboard pattern for diffuse and specular separate. Specular and diffuse lighting on Subsurface materials can appear a bit less crips.

This means the use of 32bit SceneColor (for better performance) is now possible (we might change the defaults soon).

[CL 2640696 by Martin Mittring in Main branch]