Fixed actor placement in the editor where surface snapping is enabled
The issue here is that SpawnActor has fundamentally changed such that the user-supplied transform is no longer an absolute transform, but multiplied with the class default's current transform. This affected the actor positioning code by further manipulating the desired placement transform with the default translation/rotation of the component type.

This addresses UE-19433.

[CL 2641789 by Andrew Rodham in Main branch]