Added Third Party template to the Plugin Creator Plugin
Added an option to copy the entire template directory when creating plugins as that's the easiest way of adding the Third Party library.
The copy process will modify the contents of files to change PLUGIN_NAME to the new plugin name and also ignore temporary files that may be on dev machines.

#jira UE-14890

[CL 2650909 by Matthew Griffin in Main branch]