FFindInBlueprintSearchManager now caches the asset class pointers it's already looked up
FFindInBlueprintSearchManager::OnAssetAdded was showing up as a hot-spot due to the number of times it was calling FAssetData::GetClass.

It now caches these look-ups so that it can avoid calling FindObject most of the time. This has quite the impact on asset registry processing time when dealing with large projects.

[CL 2654791 by Jamie Dale in Main branch]