Improved the caching of FPathTree (as used by FAssetRegistry)
The CachePath function of FPathTree (formerly FPathTreeNode) used to always test each component part of a path to ensure that the path was available in the tree. This not only involved always performing an FName -> FString conversion for each path that was found by the asset registry, but it also resulted in repeated work being performed when the path was already fully cached.

The new FPathTree contains a map of full parent paths to full child paths (all using FName), which allows it to early out if it's already cached the path in question. This not only avoids the repeated work, but also avoids several FName -> FString -> FName conversions that used to happen between adding a path, and then retrieving it again later (the path would previously be rebuilt from its component FString parts, and then converted back into an FName - now it just uses the FName that we already have in the map).

These changes can dramatically improve the asset registry processing performance for large projects.

[CL 2654839 by Jamie Dale in Main branch]