Making sound attenuation boundaries match the attenuation curves at min/max.
Attenuation algorithms assume that volume is 1.0 when the distance is less than the min distance and 0 when the distance is greater than the max distance. This not necessarily the case with the new custom curve feature. Now the value of attenuation will be whatever the algorithm/curve is at the min/max boundary. This won't produce a huge change for existing curves but allow for the the fact that attenuation curves may not necessarily be the canonical 1.0 -> 0.0 curve.

[CL 2658052 by Aaron McLeran in Main branch]