UE-16708
Fixed an issue where configs written to the game Agnostic saved config folder were not treated as saved configs, but defaults.
Editor Settings are now processed correctly when output to the engine saved config directory.

Note: Reverting CL# 2660460 as this was not a fix and introduced a tonne of other issues with saving arrays to default ini files.
#CodeReview Richard.Hinckley

[CL 2660672 by Terence Burns in Main branch]