Crash fix for UWorld::SendAllEndOfFrameUpdates
A TFunctionRef was being used to store some temporary lambdas, but TFunctionRef only stores a reference to the lambda (it does not keep it alive), which meant that the lambda was immediately being destroyed, and the TFunctionRef was left pointing at invalid memory. This caused a crash in non-debug builds.

#codereview Gil.Gribb

[CL 2660777 by Jamie Dale in Main branch]