Sequencer: Add default value for curves.
This gives sections a default value which is evaluated and set when there are no keys in the section. The typical workflow for setting up a transform now becomes first creating a transform track with a transform section. The current transformation of the object is set as the default value of the transform section. The default value is updated as the object is moved. But once a keyframe is set, the default value is no longer used. When all keys are removed, the default value is again in use.

Internal notes: Consolidated special flags to a KeyParams struct.

[CL 2666390 by Max Chen in Main branch]