#codereview matt.kuhlenschmidt jira: UE-12328
When the resolution is not full screen, resolution will always be smaller then the primary desktop resolution, even if the resolution is specified on the commandline(-ResX= and/or -ResY=).

In the case the primary desktop size is smaller then the minimum resolution (1280x720), the system will create a resolution of a scale of 5/6 of the desktop size in both direction, width and height. This will prevent having a window of the desktop size which conflict with the taskbar. Note that the UI layout will not support well small resolution like this.

[CL 2667816 by Alexis Matte in Main branch]