Fixed a double FText history allocation in AsNumber, AsPercent, AsCurrency, AsDate, AsTime, AsDateTime, and Format
The FText constructor taking a single FString would always allocate a default history object, which would later be replaced with the correct history object for the functions listed above.

This constructor has been changed to take an optional history object. If one is provided, the FText will just use that, otherwise it will allocate a default history object.

#codereview Justin.Sargent

[CL 2677319 by Jamie Dale in Main branch]