Added FString::Format.
Both named and ordered arguments are supported, using patterns of the form "Some text with an {Arg}" and "Some text with an {0}" respectively.
Escaped { characters can be achieved using the ` character (for exaple, "Escaped char `{}") No other escape characters are supported or necessary.
Format behaviour is similar to that of FText::Format. Numeric formatting rules may be added in future.

[CL 2680063 by Andrew Rodham in Main branch]