[AUTOMERGE]
Replaced all calls to UWorld::GetFirstPlayerController in the engine with a more readily available player controller, or UGameInstance::GetFirstLocalPlayerController, since that appears to be the actual intent. The first controller in the world's list may not be considered local, such as when a client is recording a replay.
Also changed UConsole::ConsoleCommand to use the first local player controller instead of the first player controller with a non-null player. This fixes commands like ToggleDebugCamera during client replay recording.

--------
Integrated using branch UE4-Orion-To-UE4 of change#2688431 by Ryan.Gerleve on 2015/09/11 14:24:23.

[CL 2688498 by Ryan Gerleve in Main branch]